fix: never assume env propagation
